+++ services/web/server/tests/unit/with_dbs/03/test_users.py
@@ -25,7 +25,6 @@ from faker import Faker
 from models_library.generics import Envelope
 from psycopg2 import OperationalError
 from pydantic import parse_obj_as
-from pytest_mock import MockerFixture
 from pytest_simcore.helpers.utils_assert import assert_status
 from pytest_simcore.helpers.utils_login import UserInfoDict
 from pytest_simcore.helpers.utils_tokens import (
@@ -42,7 +41,6 @@ from simcore_service_webserver._meta import API_VTAG as API_VERSION
 from simcore_service_webserver.application_settings import setup_settings
 from simcore_service_webserver.db.plugin import APP_DB_ENGINE_KEY, setup_db
 from simcore_service_webserver.groups.plugin import setup_groups
-from simcore_service_webserver.login._constants import MSG_USER_DELETED
 from simcore_service_webserver.login.plugin import setup_login
 from simcore_service_webserver.redis import (
     get_redis_user_notifications_client,
@@ -791,56 +789,3 @@ async def test_list_permissions_with_no_group_defined_returns_default_false_for_
 
     assert override_services_specifications.name == "override_services_specifications"
     assert override_services_specifications.allowed is False
-
-
-@pytest.mark.parametrize("user_role", [UserRole.USER])
-async def test_mark_account_for_deletion(
-    client: TestClient, logged_user: UserInfoDict, mocker: MockerFixture
-):
-    mock = mocker.patch(
-        "simcore_service_webserver.email._core._do_send_mail",
-        spec=True,
-    )
-
-    # is logged in
-    response = await client.get("/v0/me")
-    await assert_status(response, web.HTTPOk)
-
-    # failed check to delete account
-    response = await client.post(
-        "/v0/me:mark-deleted",
-        json={
-            "email": "WrongEmail@email.com",
-            "password": "foo",
-        },
-    )
-    await assert_status(response, web.HTTPConflict)
-
-    # success to request deletion of account
-    response = await client.post(
-        "/v0/me:mark-deleted",
-        json={
-            "email": logged_user["email"],
-            "password": logged_user["raw_password"],
-        },
-    )
-    await assert_status(response, web.HTTPOk)
-
-    # sent email?
-    mimetext = mock.call_args[1]["message"]
-    assert mimetext["Subject"]
-    assert mimetext["To"] == logged_user["email"]
-
-    # should be logged-out
-    response = await client.get("/v0/me")
-    await assert_status(response, web.HTTPUnauthorized)
-
-    # try to login again and get rejected
-    response = await client.post(
-        "/v0/auth/login",
-        json={"email": logged_user["email"], "password": logged_user["raw_password"]},
-    )
-    _, error = await assert_status(response, web.HTTPUnauthorized)
-
-    prefix_msg = MSG_USER_DELETED.format(support_email="").strip()
-    assert prefix_msg in error["errors"][0]["message"]
